[ESQL] Minor refactor for DataType (#111016)
* Remove duplicate functions from EsqlDataTypes

* Remove unused methods from DataType
