+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/expression/function/scalar/multivalue/MvConcatTests.java
@@ -16,7 +16,6 @@ import org.elasticsearch.xpack.esql.core.tree.Source;
 import org.elasticsearch.xpack.esql.core.type.DataType;
 import org.elasticsearch.xpack.esql.expression.function.AbstractScalarFunctionTestCase;
 import org.elasticsearch.xpack.esql.expression.function.TestCaseSupplier;
-import org.elasticsearch.xpack.esql.type.EsqlDataTypes;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -33,11 +32,11 @@ public class MvConcatTests extends AbstractScalarFunctionTestCase {
     public static Iterable<Object[]> parameters() {
         List<TestCaseSupplier> suppliers = new ArrayList<>();
         for (DataType fieldType : DataType.types()) {
-            if (EsqlDataTypes.isString(fieldType) == false) {
+            if (DataType.isString(fieldType) == false) {
                 continue;
             }
             for (DataType delimType : DataType.types()) {
-                if (EsqlDataTypes.isString(delimType) == false) {
+                if (DataType.isString(delimType) == false) {
                     continue;
                 }
                 for (int l = 1; l < 10; l++) {
